feat: client skeleton

+++ storage/types.go
@@ -1,13 +1,13 @@
 package storage
 
 type Message struct {
-	Offset    int64
+	Offset    uint64
 	Data      []byte `json:"data"`
 	Signature []byte `json:"signature"`
 }
 
 type Storage interface {
 	Post(message Message) error
-	GetMessages(offset int) ([]Message, error)
+	GetMessages(offset uint64) ([]Message, error)
 	Close() error
 }
